refactor(console): move asset keys assertion
