+++ src/console/src/cdn/helpers/store.rs
@@ -4,10 +4,7 @@ use crate::cdn::strategies_impls::storage::{StorageAssertions, StorageState};
 use crate::store::heap::get_controllers;
 use candid::Principal;
 use junobuild_cdn::proposals::ProposalId;
-use junobuild_cdn::storage::errors::{
-    JUNO_CDN_STORAGE_ERROR_INVALID_RELEASES_PATH, JUNO_CDN_STORAGE_ERROR_NO_PROPOSAL_FOUND,
-};
-use junobuild_shared::regex::build_regex;
+use junobuild_cdn::storage::errors::JUNO_CDN_STORAGE_ERROR_NO_PROPOSAL_FOUND;
 use junobuild_storage::store::create_batch;
 use junobuild_storage::types::interface::InitAssetKey;
 use junobuild_storage::types::runtime_state::BatchId;
@@ -26,9 +23,6 @@ pub fn init_asset_upload(
         ));
     }
 
-    // TODO: move to strategy assertion
-    assert_releases_keys(&init)?;
-
     let controllers = get_controllers();
     let config = get_config();
 
@@ -41,30 +35,4 @@ pub fn init_asset_upload(
         &StorageAssertions,
         &StorageState,
     )
-}
-
-fn assert_releases_keys(InitAssetKey { full_path, .. }: &InitAssetKey) -> Result<(), String> {
-    if full_path == "/releases/metadata.json" {
-        return Err(format!("{} is a reserved asset.", full_path).to_string());
-    }
-
-    if full_path.starts_with("/releases/satellite")
-        || full_path.starts_with("/releases/mission_control")
-        || full_path.starts_with("/releases/orbiter")
-    {
-        let re = build_regex(
-            r"^/releases/(satellite|mission_control|orbiter)-v\d+\.\d+\.\d+\.wasm\.gz$",
-        )?;
-
-        return if re.is_match(full_path) {
-            Ok(())
-        } else {
-            Err(format!(
-                "{} ({})",
-                JUNO_CDN_STORAGE_ERROR_INVALID_RELEASES_PATH, full_path
-            ))
-        };
-    }
-
-    Ok(())
 }
